pyproject.toml
[project]
name = "ai-task-copilot"
version = "0.1.0"
description = "AI Task Copilot: An agentic assistant that manages tasks across Notion, Slack, and GitHub using LLMs"
authors = [
    {name = "Siddharth Bariya",email = "sba.workspace@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "langchain-core (>=0.3.58,<0.4.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)",
    "weaviate-client (>=4.14.1,<5.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "notion-client (>=2.3.0,<3.0.0)",
    "slack-sdk (>=3.35.0,<4.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "supabase (>=2.15.1,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "respx (>=0.22.0,<0.23.0)",
    "redis (>=5.0.1,<6.0.0)",  # For Redis caching and rate limiting
    "cohere (>=5.12.0,<6.0.0)",  # For Cohere embeddings fallback
    "tenacity (>=8.2.3,<9.0.0)",  # For retry logic
    "aiohttp (>=3.9.3,<4.0.0)",  # For async HTTP requests
    "torch (>=2.2.1,<3.0.0)",  # For GPU support
    "transformers (>=4.38.2,<5.0.0)",  # Required by HuggingFace models
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langgraph-cli[inmem] (>=0.3.6,<0.4.0)",
    "langgraph (>=0.6.4,<0.7.0)",
    "langchain-google-vertexai (>=2.0.28,<3.0.0)",
    "langchain-groq (>=0.3.7,<0.4.0)",
    "tiktoken (>=0.11.0,<0.12.0)",
]

[tool.poetry]
packages = [
    { include = "app" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"
black = "^25.1.0"
flake8 = "^7.2.0"
isort = "^6.0.1"
mypy = "^1.15.0"

